Fixing np.int deprecation issue by changing it to int itself

+++ utime/sequences/base_sequence.py
@@ -199,9 +199,9 @@ class BaseSequence(_BaseSequence):
         """
         Returns:
             An ndarray of class counts across all stored SleepStudy objects
-            Shape [self.n_classes], dtype np.int
+            Shape [self.n_classes], dtype int
         """
-        counts = np.zeros(shape=[self.n_classes], dtype=np.int)
+        counts = np.zeros(shape=[self.n_classes], dtype=int)
         for im in self.dataset_queue:
             count_dict = im.get_class_counts(as_dict=True)
             for cls, count in count_dict.items():
@@ -213,7 +213,7 @@ class BaseSequence(_BaseSequence):
         """
         Returns:
             An ndarray of class frequencies comptued over all stored
-            SleepStudy objects. Shape [self.n_classes], dtype np.int
+            SleepStudy objects. Shape [self.n_classes], dtype int
         """
         counts = self.get_class_counts()
         return counts / np.sum(counts)
